Fixing various compiler warnings

+++ include/xtensor/xset_operation.hpp
@@ -178,14 +178,14 @@ namespace xt
         {
             for (size_t i = 0; i < v.size(); ++i)
             {
-                out(i) = std::lower_bound(a.cbegin(), a.cend(), v(i)) - a.cbegin();
+                out(i) = static_cast<std::size_t>(std::lower_bound(a.cbegin(), a.cend(), v(i)) - a.cbegin());
             }
         }
         else
         {
             for (size_t i = 0; i < v.size(); ++i)
             {
-                out(i) = std::upper_bound(a.cbegin(), a.cend(), v(i)) - a.cbegin();
+                out(i) = static_cast<std::size_t>(std::upper_bound(a.cbegin(), a.cend(), v(i)) - a.cbegin());
             }
         }
